fix: refine plasma one volume and move to dexs
